Reset pacList before initialize any PAC handling server
The pacList is created and filled with data at the beginning when all
the servers of all ip addresses are inited. So when the PAC file watcher
triggered a `ReSetPACProxy` action which would call the `Stop` and then
`Init` method of `PACServerHandle`, the pacList is still full of usable
yet incorrect data, thus the PAC file changes would not take place.

This fixes #1